refactor: clarify scheduler flow and test ownership
